+++ packages/api-reference/src/components/Content/Schema/Schema.vue
@@ -132,7 +132,7 @@ const handleClick = (e: MouseEvent) =>
 .schema-card {
   z-index: 0;
   position: relative;
-  font-size: var(--theme-font-size-3, var(--default-theme-font-size-3));
+  font-size: var(--theme-font-size-4, var(--default-theme-font-size-4));
   color: var(--theme-color-1, var(--default-theme-color-1));
 }