+++ packages/docusaurus/CHANGELOG.md
@@ -1,5 +1,71 @@
 # @scalar/docusaurus
 
+## 0.1.31
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.31
+
+## 0.1.30
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.30
+
+## 0.1.29
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.29
+
+## 0.1.28
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.28
+
+## 0.1.27
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.27
+
+## 0.1.26
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.26
+
+## 0.1.25
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.25
+
+## 0.1.24
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.24
+
+## 0.1.23
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.23
+
+## 0.1.22
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.22
+
+## 0.1.21
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.21
+
 ## 0.1.20
 
 ### Patch Changes