+++ packages/express-api-reference/CHANGELOG.md
@@ -1,5 +1,79 @@
 # @scalar/express-api-reference
 
+## 0.3.31
+
+### Patch Changes
+
+- Updated dependencies [ab58747]
+  - @scalar/api-reference@1.20.24
+
+## 0.3.30
+
+### Patch Changes
+
+- @scalar/api-reference@1.20.23
+
+## 0.3.29
+
+### Patch Changes
+
+- Updated dependencies [4089634]
+  - @scalar/api-reference@1.20.22
+
+## 0.3.28
+
+### Patch Changes
+
+- @scalar/api-reference@1.20.21
+
+## 0.3.27
+
+### Patch Changes
+
+- @scalar/api-reference@1.20.20
+
+## 0.3.26
+
+### Patch Changes
+
+- Updated dependencies [c99f6ff]
+  - @scalar/api-reference@1.20.18
+
+## 0.3.25
+
+### Patch Changes
+
+- Updated dependencies [5b4c8a5]
+  - @scalar/api-reference@1.20.17
+
+## 0.3.24
+
+### Patch Changes
+
+- Updated dependencies [5232eab]
+  - @scalar/api-reference@1.20.16
+
+## 0.3.23
+
+### Patch Changes
+
+- Updated dependencies [8f03345]
+  - @scalar/api-reference@1.20.15
+
+## 0.3.22
+
+### Patch Changes
+
+- Updated dependencies [78f2ab3]
+  - @scalar/api-reference@1.20.14
+
+## 0.3.21
+
+### Patch Changes
+
+- Updated dependencies [8b0691b]
+  - @scalar/api-reference@1.20.13
+
 ## 0.3.20
 
 ### Patch Changes